pyproject.toml
[project]
name = "agentic-rag-l3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.86.0",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-postgres>=0.0.17",
    "langfuse>=4.2.0",
    "langgraph>=1.1.6",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pymupdf4llm>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "redis>=7.4.0",
    "sentence-transformers>=5.4.0",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "grandalf>=0.8",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
